Better file structure for Redux This method is known as Ducks pattern. It makes code easy to access and edit. Recommended by industry experts.
