--- src/actionTypes.js
@@ -1,3 +1,0 @@
-export const BUG_ADDED = "bugAdded";
-export const BUG_REMOVED = "bugRemoved";
-export const BUG_RESOLVED = "bugResolved";
