+++ src/index.js
@@ -1,6 +1,8 @@
-import store from "./store";
+import configureStore from "./store/configureStore";
 
-import { bugAdded, bugRemoved, bugResolved } from "./actions";
+import { bugAdded, bugRemoved, bugResolved } from "./store/bugs";
+
+const store = configureStore();
 
 const unsubscribe = store.subscribe(() => {
   console.log("Store Changed!", store.getState());
