--- src/store.js
@@ -1,8 +1,0 @@
-import { createStore } from "redux";
-import { devToolsEnhancer } from "redux-devtools-extension";
-
-import reducer from "./reducers";
-
-const store = createStore(reducer, devToolsEnhancer({ trace: true }));
-
-export default store;
