+++ src/store/bugs.js
@@ -1,10 +1,31 @@
-import * as actions from "./actionTypes";
+const BUG_ADDED = "bugAdded";
+const BUG_REMOVED = "bugRemoved";
+const BUG_RESOLVED = "bugResolved";
+
+export const bugAdded = (description) => ({
+  type: BUG_ADDED,
+  payload: {
+    description,
+  },
+});
+export const bugRemoved = (id) => ({
+  type: BUG_REMOVED,
+  payload: {
+    id,
+  },
+});
+export const bugResolved = (id) => ({
+  type: BUG_RESOLVED,
+  payload: {
+    id,
+  },
+});
 
 let lastId = 0;
 
 const reducer = (state = [], action) => {
   switch (action.type) {
-    case actions.BUG_ADDED:
+    case BUG_ADDED:
       return [
         ...state,
         {
@@ -13,10 +34,10 @@ const reducer = (state = [], action) => {
           resolved: false,
         },
       ];
-    case actions.BUG_REMOVED:
+    case BUG_REMOVED:
       return state.filter((bug) => bug.id !== action.payload.id);
 
-    case actions.BUG_RESOLVED:
+    case BUG_RESOLVED:
       return state.map((bug) =>
         bug.id !== action.payload.id ? bug : { ...bug, resolved: true }
       );
