+++ src/store/configureStore.js
@@ -1,0 +1,9 @@
+import { createStore } from "redux";
+import { devToolsEnhancer } from "redux-devtools-extension";
+
+import reducer from "./bugs";
+
+export default function configureStore() {
+  const store = createStore(reducer, devToolsEnhancer({ trace: true }));
+  return store;
+};
